+++ Libraries/src/Amazon.Lambda.Annotations.SourceGenerator/Models/Attributes/AttributeModelBuilder.cs
@@ -1,4 +1,5 @@
 using System;
+using Amazon.Lambda.Annotations.ALB;
 using Amazon.Lambda.Annotations.APIGateway;
 using Amazon.Lambda.Annotations.SQS;
 using Microsoft.CodeAnalysis;
@@ -30,7 +31,7 @@ namespace Amazon.Lambda.Annotations.SourceGenerator.Models.Attributes
             else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.FromQueryAttribute), SymbolEqualityComparer.Default))
             {
                 var data = FromQueryAttributeBuilder.Build(att);
-                model = new AttributeModel<FromQueryAttribute>
+                model = new AttributeModel<APIGateway.FromQueryAttribute>
                 {
                     Data = data,
                     Type = TypeModelBuilder.Build(att.AttributeClass, context)
@@ -39,7 +40,7 @@ namespace Amazon.Lambda.Annotations.SourceGenerator.Models.Attributes
             else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.FromHeaderAttribute), SymbolEqualityComparer.Default))
             {
                 var data = FromHeaderAttributeBuilder.Build(att);
-                model = new AttributeModel<FromHeaderAttribute>
+                model = new AttributeModel<APIGateway.FromHeaderAttribute>
                 {
                     Data = data,
                     Type = TypeModelBuilder.Build(att.AttributeClass, context)
@@ -85,6 +86,60 @@ namespace Amazon.Lambda.Annotations.SourceGenerator.Models.Attributes
             {
                 var data = SQSEventAttributeBuilder.Build(att);
                 model = new AttributeModel<SQSEventAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.HttpApiAuthorizerAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = HttpApiAuthorizerAttributeBuilder.Build(att);
+                model = new AttributeModel<HttpApiAuthorizerAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.RestApiAuthorizerAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = RestApiAuthorizerAttributeBuilder.Build(att);
+                model = new AttributeModel<RestApiAuthorizerAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.ALBApiAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = ALBApiAttributeBuilder.Build(att);
+                model = new AttributeModel<ALBApiAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.ALBFromQueryAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = ALBFromQueryAttributeBuilder.Build(att);
+                model = new AttributeModel<ALB.FromQueryAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.ALBFromHeaderAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = ALBFromHeaderAttributeBuilder.Build(att);
+                model = new AttributeModel<ALB.FromHeaderAttribute>
+                {
+                    Data = data,
+                    Type = TypeModelBuilder.Build(att.AttributeClass, context)
+                };
+            }
+            else if (att.AttributeClass.Equals(context.Compilation.GetTypeByMetadataName(TypeFullNames.ALBFromBodyAttribute), SymbolEqualityComparer.Default))
+            {
+                var data = new ALB.FromBodyAttribute();
+                model = new AttributeModel<ALB.FromBodyAttribute>
                 {
                     Data = data,
                     Type = TypeModelBuilder.Build(att.AttributeClass, context)